API/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
fastapi==0.101.1
fonttools==4.42.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.2
networkx==3.1
numpy==1.25.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pydantic==2.2.0
pydantic_core==2.6.0
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
